requirements.txt
llama-cpp-python
https://huggingface.co/spacy/en_core_web_sm/resolve/main/en_core_web_sm-any-py3-none-any.whl
transformers
datasets 
torch
gradio
numpy
huggingface_hub
